+++ packages/framework/src/modules/auth/auth.constants.ts
@@ -1,4 +1,5 @@
 export enum Roles {
+    PROSPECT = 'selfservice_prospect',
     USER = 'selfservice_user',
     ADMIN = 'selfservice_admin',
 }