+++ packages/integrations/mocked/src/modules/cms/mappers/cms.login-page.mapper.ts
@@ -4,6 +4,7 @@ const LOGIN_PAGE_PL: CMS.Model.LoginPage.LoginPage = {
     title: 'Zaloguj się',
     subtitle: 'Wprowadź swój email i hasło, aby uzyskać dostęp do konta',
     username: {
+        __typename: 'Input',
         id: 'username-1',
         name: 'username',
         label: 'Nazwa użytkownika',
@@ -30,6 +31,7 @@ const LOGIN_PAGE_PL: CMS.Model.LoginPage.LoginPage = {
         ],
     },
     password: {
+        __typename: 'Input',
         id: 'password-1',
         name: 'password',
         label: 'Hasło',
@@ -93,6 +95,7 @@ const LOGIN_PAGE_EN: CMS.Model.LoginPage.LoginPage = {
     title: 'Sign in',
     subtitle: 'Please enter your email and password below to access your account.',
     username: {
+        __typename: 'Input',
         id: 'username-1',
         name: 'username',
         label: 'Username',
@@ -119,6 +122,7 @@ const LOGIN_PAGE_EN: CMS.Model.LoginPage.LoginPage = {
         ],
     },
     password: {
+        __typename: 'Input',
         id: 'password-1',
         name: 'password',
         label: 'Password',
@@ -203,6 +207,7 @@ const LOGIN_PAGE_DE: CMS.Model.LoginPage.LoginPage = {
     title: 'Einloggen',
     subtitle: 'Geben Sie Ihre E-Mail und Ihr Passwort ein, um auf Ihr Konto zuzugreifen',
     username: {
+        __typename: 'Input',
         id: 'username-1',
         name: 'username',
         label: 'Benutzername',
@@ -229,6 +234,7 @@ const LOGIN_PAGE_DE: CMS.Model.LoginPage.LoginPage = {
         ],
     },
     password: {
+        __typename: 'Input',
         id: 'password-1',
         name: 'password',
         label: 'Passwort',