+++ apps/api-harmonization/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @o2s/api-harmonization
 
+## 1.8.0
+
+### Minor Changes
+
+- e81621c: added `PROSPECT` role for freshly-created users
+
+### Patch Changes
+
+- Updated dependencies [e81621c]
+    - @o2s/integrations.mocked@1.9.0
+    - @o2s/framework@1.9.0
+
+## 1.7.0
+
+### Minor Changes
+
+- fb09383: Removing unused import
+
+### Patch Changes
+
+- Updated dependencies [fb09383]
+- Updated dependencies [fb09383]
+    - @o2s/integrations.mocked@1.8.0
+    - @o2s/framework@1.8.0
+
+## 1.6.1
+
+### Patch Changes
+
+- ca33e9a: fix: fixed api-harmonization generator
+
 ## 1.6.0
 
 ### Minor Changes